ACA-68 - Refactor request notice of change out to its own service
